backend/requirements.txt
djangorestframework
pyyaml
django-cors-headers
